+++ lib/appfuel/storage/web_api.rb
@@ -1,0 +1,2 @@
+require_relative 'web_api/repository'
+require_relative 'web_api/http_model'